Fixed processor model reported in #123. Implementation of #156.
The bug related to the computation of the processor model was fixed (see
comment in #123).

As part of the linting process (#156) the code structure was slightly
changed in order to delegate the implementation of the Processor
operating system functionality to a .cpp unit. This improves the
consistency with the other operating system classes.