#197 Implemented the construction of an AnyType from an Object.
Finished the implementation of AnyType from Object. With this method it
now possible to generate an AnyType from any class that has been
registered in the ClassRegistryDatabase. Consequently, this has required
supporting the registration of typeid(class).name() in the
ClassRegistryDatabase.

As part of this development, BitSize was renamed to GetNumberOfBits in
BitBoolean, BitRange and FractionalInteger.

There were missing unit tests in AnyType which were also created.