#225 Implementation of GlobalObjectsDatabase
The implementation of the ClassRegistryDatabase with a StaticList
backend has generated a racing condition between the HeapDatabase and
ClassRegistryDatabase singletons. This was making the test program crash
at the end of its execution. As a consequence it was possible to run the
tests for this story.

To address the issue above a global object manager, named
GlobalObjectsDatabase, was implemented. This will be the only true
global object in the framework. All the other singleton objects will be
dynamically allocated in the heap and registered in this
GlobalObjectDatabase.