#276 solved regression bug in IOBufferTest.cpp
In order to increase the coverage of the StreamStringIOBuffer Printf
function, a test with a very long string was added. Given that the same
string was also used by the IOBufferTest this has caused one of the
tests in IOBuffer to fail (given that the string was larger than the
buffer size). The solution was to change the IOBufferTest to compare the
string only up to its BufferSize (as the rest will be correctly
truncated).