Bug solved in Threads::BeginThread.
There was a racing condition between the thread creation (performed in
BeginThread) and the actual start of the thread which happens in
SystemThreadFunction (see threadInfo->UserThreadFunction()). If a user
called Thread::IsAlive (or any other method that required the Thread to
be already added to the database) before ThreadsDatabase::NewEntry was
called (which was being done in the context of the new thread) this
would lead to errors.

Now ThreadsDatabase::NewEntry is called in the context of who called
Threads::BeginThread so that consistency of the database is guaranteed.
That is, Threads::BeginThread does not return before: 1) the thread is
really started and 2) the thread is added to the database.