#GAMDataSource memory optimisation
The memory backend of the GAMDataSource was optimised to use a single
block of memory for the data source and to use an indexer array to find
the signals. This was mostly driven by the fact that we have decided to
use single buffer implementation.

The reason to use a single buffer implementation was that when the
defautl values are being loaded, it means that the variable was not
being used in the current state and thus there is no need to protect
this memory area with another buffer.