Solve bug on MemoryMapAsyncTriggerOutputBroker
The MemoryMapAsyncTriggerOutputBroker was wrongly assuming that the
first signal of the GAM writing to the broker was the Trigger signal.
This not necessarily true. The only requirement is that the first signal
on the DataSource must be the Trigger signal. But the first signal in
the DataSource does not have to be necessarily connected to the first
signal of the GAM. 

The Broker was corrected, linted and tested.