added FromFormattedStreamsConversionFactory.
The conversion between StructuredDataI and they charcter based
representation is now fully integrated.
All characterStream types are supported as well as new
FormattedCharacterStreams
The distinction is only in the TypeDescription itself, where it is now
possible to specify the format using a 4 char notation: CDB,XML,JSON
CDB is the default format if none is specified.